+++ jobs/ml_jobs/algo_training/similar_offers/deploy_model.py
@@ -50,7 +50,9 @@ def download_model(artifact_uri):
     tf_reco = tf.keras.models.load_model("./model/")
     item_list = tf_reco.item_layer.layers[0].get_vocabulary()
     model_weights = tf_reco.item_layer.layers[1].get_weights()[0]
-    np.save("./metadata/weights.npy", model_weights, allow_pickle=True)
+    np.save(
+        "./metadata/weights.npy", model_weights.astype(np.float32), allow_pickle=True
+    )
     np.save("./metadata/items.npy", item_list, allow_pickle=True)
